Update WinUI3 client, isolate most shared concepts to the new Shared project
