+++ PathFinding.Shared/Annotations.cs
