+++ PathFinding.Shared/Models/ConveyorTile.cs
@@ -1,0 +1,11 @@
+﻿namespace PathFinding.Models;
+
+public class ConveyorTile
+{
+    public (int X, int Y) Direction;
+    public Tile Tile;
+    public List<Item> Items = new();
+    public Conveyor Conveyor;
+    public ConveyorTile NextConveyorTile;
+}
+
