+++ PathFinding.Shared/Models/State.cs
@@ -1,0 +1,43 @@
+﻿using PathFinding.Persistence;
+
+namespace PathFinding.Models;
+
+public class State
+{
+    public int X;
+    public int Y;
+    private readonly IStatePersistence _sp;
+    private Tile[,] _tileGrid;
+    public List<Tile> Tiles { get; set; } = new();
+    public int TileSize { get; }
+    public Tile[,] TileGrid { get => _tileGrid; set => _tileGrid = value; }
+
+    public State(int x, int y, int tileSize, IStatePersistence sp)
+    {
+        X = x;
+        Y = y;
+        _sp = sp;
+        TileSize = tileSize;
+        TileGrid = new Tile[x, y];
+        SetTiles();
+    }
+
+    private void SetTiles()
+    {
+        var i = 0;
+        for (var x = 0; x < X; x++)
+        {
+            for (var y = 0; y < Y; y++)
+            {
+                var tempTile = new Tile(x, y, true, i++);
+                TileGrid[x, y] = tempTile;
+                Tiles.Add(tempTile);
+            }
+        }
+    }
+
+    public string SetupMapString() => _sp.SetupMapString(ref X, ref Y, ref _tileGrid);
+
+    public (int X, int Y, string[] mapStringArray, bool Success) UploadMapString(string mapString) => _sp.UploadMapString(mapString);
+}
+
