+++ PathFinding.WinUI/App.xaml.cs
@@ -16,10 +16,8 @@ using Windows.ApplicationModel.Activation;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
 using Microsoft.Extensions.DependencyInjection;
-using PathFinding.Stuff;
-
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
+using PathFinding.Persistence;
+using PathFinding.ViewModels;
 
 namespace Pathfinding.WinUI;
 /// <summary>
