+++ Picking.sln
@@ -12,7 +12,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PathFinding.WinUI", "PathFinding.WinUI\PathFinding.WinUI.csproj", "{F3FB5FC2-0555-4B9B-8D57-3EE04CA43956}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PathFinding.WinUI", "PathFinding.WinUI\PathFinding.WinUI.csproj", "{F3FB5FC2-0555-4B9B-8D57-3EE04CA43956}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PathFinding.Shared", "PathFinding.Shared\PathFinding.Shared.csproj", "{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -82,6 +84,22 @@ Global
 		{F3FB5FC2-0555-4B9B-8D57-3EE04CA43956}.Release|x86.ActiveCfg = Release|x86
 		{F3FB5FC2-0555-4B9B-8D57-3EE04CA43956}.Release|x86.Build.0 = Release|x86
 		{F3FB5FC2-0555-4B9B-8D57-3EE04CA43956}.Release|x86.Deploy.0 = Release|x86
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Debug|arm64.ActiveCfg = Debug|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Debug|arm64.Build.0 = Debug|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Debug|x64.Build.0 = Debug|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Debug|x86.Build.0 = Debug|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Release|arm64.ActiveCfg = Release|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Release|arm64.Build.0 = Release|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Release|x64.ActiveCfg = Release|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Release|x64.Build.0 = Release|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Release|x86.ActiveCfg = Release|Any CPU
+		{6B6A0C74-4E3B-41AA-B916-AE2B63773E88}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
